requirements.txt
python-dotenv
pydantic
SQLAlchemy
pandas
bs4
glom
psycopg2-binary
scrapy
zendriver
streamlit
